refactored messages

+++ src/consts.rs
@@ -1,5 +1,5 @@
 use std::time::Duration;
 
 pub const TIMEOUT: Duration = Duration::from_secs(10);
-pub const BLOCKSIZE: usize = 1 << 14;
-pub const MAXREQUESTS: usize = 5;
+pub const BLOCKSIZE: u32 = 1 << 14;
+pub const MAXREQUESTS: u32 = 5;
